backend/requirements.txt
fastapi
uvicorn
paho-mqtt
sqlalchemy
psycopg2-binary
scikit-learn
joblib
numpy
